refactor: replace menu
