+++ apps/client/src/features/editors/Editor.tsx
@@ -1,10 +1,10 @@
-import { lazy, useEffect } from 'react';
+import { lazy, useCallback, useEffect } from 'react';
 
-import ErrorBoundary from '../../common/components/error-boundary/ErrorBoundary';
+import ProductionNavigationMenu from '../../common/components/navigation-menu/ProductionNavigationMenu';
+import useElectronEvent from '../../common/hooks/useElectronEvent';
 import AppSettings from '../app-settings/AppSettings';
 import { SettingsOptionId } from '../app-settings/settingsStore';
 import useAppSettingsNavigation from '../app-settings/useAppSettingsNavigation';
-import MenuBar from '../menu/MenuBar';
 import Overview from '../overview/Overview';
 
 import styles from './Editor.module.scss';
@@ -15,14 +15,49 @@ const MessageControl = lazy(() => import('../control/message/MessageControlExpor
 
 export default function Editor() {
   const { isOpen, setLocation, close } = useAppSettingsNavigation();
+  const { isElectron } = useElectronEvent();
 
-  const handleSettings = (newTab?: SettingsOptionId) => {
-    if (isOpen) {
-      close();
-    } else {
-      setLocation(newTab ?? 'project');
+  const handleSettings = useCallback(
+    (newTab?: SettingsOptionId) => {
+      if (isOpen) {
+        close();
+      } else {
+        setLocation(newTab ?? 'project');
+      }
+    },
+    [close, isOpen, setLocation],
+  );
+
+  // Handle keyboard shortcuts
+  const handleKeyPress = useCallback(
+    (event: KeyboardEvent) => {
+      // handle held key
+      if (event.repeat) return;
+
+      // check if the ctrl key is pressed
+      if (event.ctrlKey || event.metaKey) {
+        // ctrl + , (settings)
+        if (event.key === ',') {
+          handleSettings();
+          event.preventDefault();
+          event.stopPropagation();
+        }
+      }
+    },
+    [handleSettings],
+  );
+
+  // register ctrl + , to open settings
+  useEffect(() => {
+    if (isElectron) {
+      document.addEventListener('keydown', handleKeyPress);
     }
-  };
+    return () => {
+      if (isElectron) {
+        document.removeEventListener('keydown', handleKeyPress);
+      }
+    };
+  }, [handleKeyPress, isElectron]);
 
   // Set window title
   useEffect(() => {
@@ -31,9 +66,7 @@ export default function Editor() {
 
   return (
     <div className={styles.mainContainer} data-testid='event-editor'>
-      <ErrorBoundary>
-        <MenuBar openSettings={handleSettings} isSettingsOpen={isOpen} />
-      </ErrorBoundary>
+      <ProductionNavigationMenu handleSettings={handleSettings} />
       {isOpen ? (
         <AppSettings />
       ) : (
