--- apps/client/src/features/menu/MenuBar.module.scss
@@ -1,20 +1,0 @@
-.menu {
-  grid-area: menu;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 0.5em;
-  padding-bottom: 1rem
-}
-
-.gap {
-  height: 1em;
-}
-
-.open {
-  background: $blue-700;
-}
-
-.bottom {
-  margin-top: auto;
-}
